refactor(core): move edge prop types
