+++ packages/core/src/components/Edges/SmoothStepEdge.ts
@@ -1,12 +1,7 @@
 import type { FunctionalComponent } from 'vue'
 import BaseEdge from './BaseEdge'
-import type { BaseEdgeProps, EdgePositions, EdgeProps, SmoothStepPathOptions } from '~/types'
+import type { SmoothStepEdgeProps } from '~/types'
 import { Position } from '~/types'
-
-export type SmoothStepEdgeProps = EdgePositions &
-  Omit<BaseEdgeProps, 'labelX' | 'labelY' | 'path'> &
-  Pick<EdgeProps, 'sourcePosition' | 'targetPosition'> &
-  SmoothStepPathOptions
 
 const SmoothStepEdge: FunctionalComponent<SmoothStepEdgeProps> = function (
   { sourcePosition = Position.Bottom, targetPosition = Position.Top, ...props },
