+++ packages/core/src/components/Edges/index.ts
@@ -1,9 +1,9 @@
 export { default as BaseEdge } from './BaseEdge'
-export { default as BezierEdge, BezierEdgeProps } from './BezierEdge'
-export { default as SimpleBezierEdge, SimpleBezierEdgeProps } from './SimpleBezierEdge'
-export { default as StepEdge, StepEdgeProps } from './StepEdge'
-export { default as SmoothStepEdge, SmoothStepEdgeProps } from './SmoothStepEdge'
-export { default as StraightEdge, StraightEdgeProps } from './StraightEdge'
+export { default as BezierEdge } from './BezierEdge'
+export { default as SimpleBezierEdge } from './SimpleBezierEdge'
+export { default as StepEdge } from './StepEdge'
+export { default as SmoothStepEdge } from './SmoothStepEdge'
+export { default as StraightEdge } from './StraightEdge'
 export { default as EdgeAnchor } from './EdgeAnchor'
 export { default as EdgeText } from './EdgeText.vue'
 export { default as EdgeWrapper } from './EdgeWrapper'
